requirements.txt
# ResearchPulse - Python Dependencies
# Install with: pip install -r requirements.txt

# =============================================================================
# Core Framework
# =============================================================================
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# =============================================================================
# LangChain - ReAct Agent and Tools
# =============================================================================
langchain>=0.1.0,<1.0.0
langchain-core>=0.1.0,<1.0.0
langchain-community>=0.0.10,<1.0.0
langchain-openai>=0.0.5,<1.0.0

# =============================================================================
# Pinecone - Vector Store for RAG
# =============================================================================
pinecone-client>=3.0.0,<4.0.0
langchain-pinecone>=0.0.3,<1.0.0

# =============================================================================
# Supabase - Primary Database (per Course Project requirements)
# =============================================================================
supabase>=2.0.0,<3.0.0

# =============================================================================
# Embeddings and ML
# =============================================================================
openai>=1.10.0,<2.0.0
tiktoken>=0.5.0,<1.0.0

# =============================================================================
# arXiv Integration
# =============================================================================
arxiv>=2.1.0,<3.0.0
feedparser>=6.0.0,<7.0.0

# =============================================================================
# HTTP and Async
# =============================================================================
httpx>=0.26.0,<1.0.0
aiofiles>=23.2.0,<24.0.0

# =============================================================================
# Data Handling
# =============================================================================
python-dateutil>=2.8.0,<3.0.0

# =============================================================================
# Development and Testing (optional)
# =============================================================================
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.23.0,<1.0.0
black>=24.1.0,<25.0.0
isort>=5.13.0,<6.0.0
mypy>=1.8.0,<2.0.0
